+++ src/xmat-models/xmat-confirm-dialog.model.ts
@@ -1,9 +1,11 @@
 import {XmatDialogContentComponent} from "../xmat-dialog/ts/xmat-dialog-content.component";
+import {ThemePalette} from "@angular/material";
 
 export interface XmatConfirmDialogData {
-    cancelText: string;
-    confirmText: string;
+    cancelText?: string;
+    confirmText?: string;
     dialogContent: string | XmatDialogContentComponent;
-    hideCancelButton: boolean;
+    hideCancelButton?: boolean;
+    confirmColor?: ThemePalette;
     title: string;
 }